refactor: move icrc constants
