+++ src/constants/icrc.constants.ts
@@ -1,0 +1,13 @@
+export const ICRC25_REQUEST_PERMISSIONS = 'icrc25_request_permissions';
+export const ICRC25_PERMISSIONS = 'icrc25_permissions';
+export const ICRC25_SUPPORTED_STANDARDS = 'icrc25_supported_standards';
+// TODO: ICRC27 to be defined
+export const ICRC27_ACCOUNTS = 'icrc27_accounts';
+export const ICRC29_STATUS = 'icrc29_status';
+
+export const ICRC25_PERMISSION_GRANTED = 'granted';
+export const ICRC25_PERMISSION_DENIED = 'denied';
+export const ICRC25_PERMISSION_ASK_ON_USE = 'ask_on_user';
+
+export const ICRC25 = 'ICRC-25';
+export const ICRC29 = 'ICRC-29';
